refactor(relays): split relays.ts into relays/ modules
